Show "Signing in" screen for desktop-forced passkey login
When the desktop app forces a passkey login it opens the login page in the
system browser, where the passkey ceremony is triggered automatically. Show
the same "Signing in" screen used during desktop redirect instead of the
login buttons, since no manual interaction is needed. The passkey provider
stays mounted to drive the ceremony and reveals a retry button if it fails.
