+++ app/scenes/Login/Login.tsx
@@ -40,6 +40,8 @@ import { BackButton } from "./components/BackButton";
 import { Background } from "./components/Background";
 import { Centered } from "./components/Centered";
 import { Notices } from "./components/Notices";
+import { PasskeyAuthenticationProvider } from "./components/PasskeyAuthenticationProvider";
+import { SigningIn } from "./components/SigningIn";
 import { SwitchHostButton } from "./components/SwitchHostButton";
 import { navigateToSubdomain } from "./urls";
 import lazyWithRetry from "~/utils/lazyWithRetry";
@@ -122,6 +124,14 @@ function Login({ children, onBack }: Props) {
   // ceremony even when this browser already has a session.
   const isPasskeyLogin = query.get("method") === "passkey";
 
+  // When the desktop app forces a passkey login it opens this page in the
+  // system browser, where the ceremony is triggered automatically. Show the
+  // "Signing in" screen rather than the login buttons.
+  const isDesktopPasskeyRedirect =
+    isPasskeyLogin &&
+    query.get("client") === Client.Desktop &&
+    !Desktop.isElectron();
+
   if (auth.authenticated && !isPasskeyLogin) {
     const postLoginPath = spendPostLoginPath();
     if (postLoginPath) {
@@ -167,6 +177,22 @@ function Login({ children, onBack }: Props) {
   // indicator here that's delayed by 250ms
   if (!config) {
     return <LoadingIndicator />;
+  }
+
+  // The passkey ceremony is triggered automatically here, so render the
+  // "Signing in" screen in place of the login buttons. The passkey provider is
+  // still mounted (hidden) to drive the ceremony and form submission, revealing
+  // a retry button if the ceremony fails.
+  if (isDesktopPasskeyRedirect) {
+    return (
+      <Background>
+        <ChangeLanguage locale={detectLanguage()} />
+        <Centered gap={12}>
+          <SigningIn />
+          <PasskeyAuthenticationProvider />
+        </Centered>
+      </Background>
+    );
   }
 
   const isCustomDomain = parseDomain(window.location.origin).custom;
